feat: add `queryStream` and `queryNodeStream` methods to MySQL, PG & SQLite (#47)
This adds a proper documented method for querying a database and streaming the results out. This can be used for querying very large tables when you want to fetch all results and process them as a stream.

The SQLite database only implements `queryStream`, which returns an async iterable. Postgres and MySQL implement both methods, and also correctly implement back-pressure.