fix: prevent crash when idle connection is terminated
In rare circumstances, an error could be thrown on a connection that is not fully idle (i.e. it's within a transaction/task) but does not have an active query. We were previously not treating these errors as idle connection errors, or handling them to throw with the next query. The result was that these errors would just crash the node.js process.

This change should fix that behaviour.